+++ packages/genesys/packages/react-components/src/Dialog/Dialog.types.ts
@@ -2,10 +2,11 @@ import { DialogButton, DialogProps as CoreDialogProps } from "@peersyst/react-co
 import { DIALOG_SIZES } from "./Dialog.constants";
 import { Demand } from "@peersyst/react-types";
 import { ButtonProps } from "../Button";
+import { ComponentsConfig } from "../config";
 
 export type DialogSize = typeof DIALOG_SIZES[number];
 
-export interface DialogProps extends CoreDialogProps {
+export interface DialogProps extends CoreDialogProps<ComponentsConfig["Dialog"]> {
     size?: DialogSize | string;
 }